@@ -2,24 +2,23 @@ import 'package:flutter/material.dart';
 
 // Default values from the Flutter's TabBar.
 const double _kTabHeight = 46.0;
-const double _kTextAndIconTabHeight = 72.0;
 
 class ButtonsTabBar extends StatefulWidget implements PreferredSizeWidget {
   ButtonsTabBar({
-    Key key,
-    @required this.tabs,
+    Key? key,
+    required this.tabs,
     this.controller,
     this.duration = 250,
-    this.backgroundColor = Colors.blueAccent,
-    this.unselectedBackgroundColor = Colors.grey,
+    this.backgroundColor,
+    this.unselectedBackgroundColor,
     this.labelStyle,
     this.unselectedLabelStyle,
-    this.borderWidth,
+    this.borderWidth = 0,
     this.borderColor = Colors.black,
     this.unselectedBorderColor,
-    this.physics,
-    this.contentPadding,
-    this.buttonMargin,
+    this.physics = const BouncingScrollPhysics(),
+    this.contentPadding = const EdgeInsets.symmetric(horizontal: 4),
+    this.buttonMargin = const EdgeInsets.all(4),
     this.labelSpacing = 4.0,
     this.radius = 7.0,
     this.height = _kTabHeight,
@@ -35,46 +34,49 @@ class ButtonsTabBar extends StatefulWidget implements PreferredSizeWidget {
   ///
   /// If [TabController] is not provided, then the value of [DefaultTabController.of]
   /// will be used.
-  final TabController controller;
+  final TabController? controller;
 
   /// The duration in milliseconds of the transition animation.
   final int duration;
 
   /// The background [Color] of the button on its selected state.
-  final Color backgroundColor;
+  ///
+  /// If [Color] is not provided, [Theme.of(context).accentColor] is used.
+  final Color? backgroundColor;
 
   /// The background [Color] of the button on its unselected state.
-  final Color unselectedBackgroundColor;
+  ///
+  /// If [Color] is not provided, [Colors.grey[300]] is used.
+  final Color? unselectedBackgroundColor;
 
   /// The [TextStyle] of the button's [Text] on its selected state. The color provided
   /// on the TextStyle will be used for the [Icon]'s color.
   ///
   /// The default value is: [TextStyle(color: Colors.white)].
-  final TextStyle labelStyle;
+  final TextStyle? labelStyle;
 
   /// The [TextStyle] of the button's [Text] on its unselected state. The color provided
   /// on the TextStyle will be used for the [Icon]'s color.
   ///
   /// The default value is: [TextStyle(color: Colors.black)].
-  final TextStyle unselectedLabelStyle;
+  final TextStyle? unselectedLabelStyle;
 
   /// The with of solid [Border] for each button. If no value is provided, the border
   /// is not drawn.
   ///
-  /// The default value is: null.
+  /// The default value is: 0.
   final double borderWidth;
 
-  /// The [Color] of solid [Border] for each button. To hide the [Border], provide
-  /// [Colors.transparent].
+  /// The [Color] of solid [Border] for each button.
   ///
   /// The default value is: [Colors.black].
   final Color borderColor;
 
   /// The [Color] of solid [Border] for each button. If no value is provided, the value of
-  /// [this.borderColor] is used. To hide the [Border], provide [Colors.transparent].
+  /// [this.borderColor] is used.
   ///
   /// The default value is: [Colors.black].
-  final Color unselectedBorderColor;
+  final Color? unselectedBorderColor;
 
   /// The physics used for the [ScrollController] of the tabs list.
   ///
@@ -83,7 +85,7 @@ class ButtonsTabBar extends StatefulWidget implements PreferredSizeWidget {
 
   /// The [EdgeInsets] used for the [Padding] of the buttons' content.
   ///
-  /// The default value is [EdgeInsets.all(4)].
+  /// The default value is [EdgeInsets.symmetric(horizontal: 4)].
   final EdgeInsets contentPadding;
 
   /// The [EdgeInsets] used for the [Margin] of the buttons.
@@ -98,21 +100,17 @@ class ButtonsTabBar extends StatefulWidget implements PreferredSizeWidget {
   /// The value of the [BorderRadius.circular] applied to each button.
   final double radius;
 
-  /// Override the default height. If no value is provided, the material height, 46.0,
-  /// is used.
-  final double height;
+  /// Override the default height.
+  ///
+  /// If no value is provided, the material height, 46.0, is used. If height is [null],
+  /// the height is computed by summing the material height, 46, and the vertical values
+  /// for [contentPadding] and [buttonMargin].
+  final double? height;
 
   @override
   Size get preferredSize {
-    for (Widget item in tabs) {
-      if (item is Tab) {
-        final Tab tab = item;
-        if (tab.text != null && tab.icon != null)
-          return Size.fromHeight(
-              height != _kTabHeight ? height : _kTextAndIconTabHeight);
-      }
-    }
-    return Size.fromHeight(height);
+    return Size.fromHeight(height ??
+        (_kTabHeight + contentPadding.vertical + buttonMargin.vertical));
   }
 
   @override
@@ -121,35 +119,14 @@ class ButtonsTabBar extends StatefulWidget implements PreferredSizeWidget {
 
 class _ButtonsTabBarState extends State<ButtonsTabBar>
     with TickerProviderStateMixin {
-  TabController _controller;
+  TabController? _controller;
 
   ScrollController _scrollController = new ScrollController();
-  ScrollPhysics _scrollPhysics;
 
-  AnimationController _animationController;
+  late AnimationController _animationController;
 
-  List<GlobalKey> _tabKeys;
+  late List<GlobalKey> _tabKeys;
   GlobalKey _tabsContainerKey = GlobalKey();
-
-  Animation<Color> _colorTweenForegroundActivate;
-  Animation<Color> _colorTweenForegroundDeactivate;
-  Animation<Color> _colorTweenBackgroundActivate;
-  Animation<Color> _colorTweenBackgroundDeactivate;
-  Animation<Color> _colorTweenBorderActivate;
-  Animation<Color> _colorTweenBorderDeactivate;
-
-  Color _unselectedForegroundColor;
-  Color _foregroundColor;
-
-  Color _unselectedBorderColor;
-  Color _borderColor;
-  bool _borderAnimation;
-
-  TextStyle _unselectedLabelStyle;
-  TextStyle _labelStyle;
-
-  EdgeInsets _contentPadding;
-  EdgeInsets _buttonMargin;
 
   int _currentIndex = 0;
   int _prevIndex = -1;
@@ -157,7 +134,7 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
   double _prevAniValue = 0;
 
   // check the direction of the text LTR or RTL
-  bool _textLTR;
+  late bool _textLTR;
 
   @override
   void initState() {
@@ -165,54 +142,18 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
 
     _tabKeys = widget.tabs.map((Widget tab) => GlobalKey()).toList();
 
-    _scrollPhysics = widget.physics ?? BouncingScrollPhysics();
-
-    _unselectedLabelStyle =
-        widget.unselectedLabelStyle ?? TextStyle(color: Colors.black);
-    _labelStyle = widget.labelStyle ?? TextStyle(color: Colors.white);
-
-    _foregroundColor = _labelStyle.color ?? Colors.white;
-    _unselectedForegroundColor = _unselectedLabelStyle.color ?? Colors.black;
-
-    _borderColor = widget.borderColor;
-    _unselectedBorderColor = widget.unselectedBorderColor ?? widget.borderColor;
-    _borderAnimation = _borderColor != _unselectedBorderColor;
-
-    _contentPadding = widget.contentPadding ?? EdgeInsets.all(4);
-    _buttonMargin = widget.buttonMargin ?? EdgeInsets.all(4);
-
     _animationController = AnimationController(
         vsync: this, duration: Duration(milliseconds: widget.duration));
-    _colorTweenBackgroundActivate = ColorTween(
-            begin: widget.unselectedBackgroundColor,
-            end: widget.backgroundColor)
-        .animate(_animationController);
-    _colorTweenBackgroundDeactivate = ColorTween(
-            begin: widget.backgroundColor,
-            end: widget.unselectedBackgroundColor)
-        .animate(_animationController);
-    _colorTweenForegroundActivate =
-        ColorTween(begin: _unselectedForegroundColor, end: _foregroundColor)
-            .animate(_animationController);
-    _colorTweenForegroundDeactivate =
-        ColorTween(begin: _foregroundColor, end: _unselectedForegroundColor)
-            .animate(_animationController);
-
-    if (_borderAnimation) {
-      _colorTweenBorderActivate =
-          ColorTween(begin: _unselectedBorderColor, end: _borderColor)
-              .animate(_animationController);
-      _colorTweenBorderDeactivate =
-          ColorTween(begin: _borderColor, end: _unselectedBorderColor)
-              .animate(_animationController);
-    }
 
     // so the buttons start in their "final" state (color)
     _animationController.value = 1.0;
+    _animationController.addListener(() {
+      setState(() {});
+    });
   }
 
   void _updateTabController() {
-    final TabController newController =
+    final TabController? newController =
         widget.controller ?? DefaultTabController.of(context);
     assert(() {
       if (newController == null) {
@@ -228,15 +169,13 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
     if (newController == _controller) return;
 
     if (_controllerIsValid) {
-      _controller.animation.removeListener(_handleTabAnimation);
-      _controller.removeListener(_handleController);
+      _controller?.animation!.removeListener(_handleTabAnimation);
+      _controller?.removeListener(_handleController);
     }
     _controller = newController;
-    if (_controller != null) {
-      _controller.animation.addListener(_handleTabAnimation);
-      _controller.addListener(_handleController);
-      _currentIndex = _controller.index;
-    }
+    _controller?.animation!.addListener(_handleTabAnimation);
+    _controller?.addListener(_handleController);
+    _currentIndex = _controller!.index;
   }
 
   // If the TabBar is rebuilt with a new tab controller, the caller should
@@ -258,13 +197,6 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
       _updateTabController();
     }
 
-    // Update the style
-    if (widget.labelStyle != _labelStyle)
-      _labelStyle = widget.labelStyle ?? TextStyle(color: Colors.white);
-    if (widget.unselectedLabelStyle != _unselectedLabelStyle)
-      _unselectedLabelStyle =
-          widget.unselectedLabelStyle ?? TextStyle(color: Colors.black);
-
     if (widget.tabs.length > oldWidget.tabs.length) {
       final int delta = widget.tabs.length - oldWidget.tabs.length;
       _tabKeys.addAll(List<GlobalKey>.generate(delta, (int n) => GlobalKey()));
@@ -274,76 +206,76 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
   }
 
   void _handleController() {
-    if (_controller.indexIsChanging) {
+    if (_controller!.indexIsChanging) {
       // update highlighted index when controller index is changing
-      _goToIndex(_controller.index);
+      _goToIndex(_controller!.index);
     }
   }
 
   @override
   void dispose() {
     if (_controllerIsValid) {
-      _controller.animation.removeListener(_handleTabAnimation);
-      _controller.removeListener(_handleController);
+      _controller!.animation!.removeListener(_handleTabAnimation);
+      _controller!.removeListener(_handleController);
     }
     _controller = null;
-    _scrollController?.dispose();
+    _scrollController.dispose();
     super.dispose();
   }
 
-  Widget _buildButton(int index, Tab tab) {
-    final TextStyle textStyle = (index == _currentIndex
-        ? TextStyle.lerp(
-            _unselectedLabelStyle, _labelStyle, _animationController.value)
-        : (index == _prevIndex
-            ? TextStyle.lerp(
-                _labelStyle, _unselectedLabelStyle, _animationController.value)
-            : _unselectedLabelStyle));
+  Widget _buildButton(
+    int index,
+    Tab tab,
+  ) {
+    final double animationValue;
+    if (index == _currentIndex)
+      animationValue = _animationController.value;
+    else if (index == _prevIndex)
+      animationValue = 1 - _animationController.value;
+    else
+      animationValue = 0;
 
-    return Container(
+    final TextStyle? textStyle = TextStyle.lerp(
+        widget.unselectedLabelStyle ?? TextStyle(color: Colors.black),
+        widget.labelStyle ?? TextStyle(color: Colors.white),
+        animationValue);
+    final Color? backgroundColor = Color.lerp(
+        widget.unselectedBackgroundColor ?? Colors.grey[300],
+        widget.backgroundColor ?? Theme.of(context).accentColor,
+        animationValue);
+    final Color? borderColor = Color.lerp(
+        widget.unselectedBorderColor ?? widget.borderColor,
+        widget.borderColor,
+        animationValue);
+    final Color foregroundColor = textStyle?.color ?? Colors.black;
+
+    return Padding(
       key: _tabKeys[index],
       // padding for the buttons
-      margin: _buttonMargin,
-      child: FlatButton(
-        padding: _contentPadding,
-        materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        // get the color of the button's background (dependent of its state)
-        color: index == _currentIndex
-            ? _colorTweenBackgroundActivate.value
-            : (index == _prevIndex
-                ? _colorTweenBackgroundDeactivate.value
-                : widget.unselectedBackgroundColor),
-        // make the button a rectangle with round corners
-        shape: RoundedRectangleBorder(
-          side: widget.borderWidth == null
-              ? BorderSide.none
-              : BorderSide(
-                  color: _borderAnimation
-                      ? (index == _currentIndex
-                          ? _colorTweenBorderActivate.value
-                          : (index == _prevIndex
-                              ? _colorTweenBorderDeactivate.value
-                              : _unselectedBorderColor))
-                      : _borderColor,
-                  width: widget.borderWidth,
-                  style: BorderStyle.solid),
-          borderRadius: BorderRadius.circular(widget.radius),
+      padding: widget.buttonMargin,
+      child: TextButton(
+        onPressed: () => _controller?.animateTo(index),
+        style: TextButton.styleFrom(
+          padding: widget.contentPadding,
+          backgroundColor: backgroundColor,
+          textStyle: textStyle,
+          tapTargetSize: MaterialTapTargetSize.shrinkWrap,
+          shape: RoundedRectangleBorder(
+            side: (widget.borderWidth == 0)
+                ? BorderSide.none
+                : BorderSide(
+                    color: borderColor ?? Colors.black,
+                    width: widget.borderWidth,
+                    style: BorderStyle.solid),
+            borderRadius: BorderRadius.circular(widget.radius),
+          ),
         ),
-        onPressed: () {
-          _controller?.animateTo(index);
-        },
         child: Row(
           children: <Widget>[
             tab.icon != null
                 ? IconTheme.merge(
-                    data: IconThemeData(
-                        size: 24.0,
-                        color: index == _currentIndex
-                            ? _colorTweenForegroundActivate.value
-                            : (index == _prevIndex
-                                ? _colorTweenForegroundDeactivate.value
-                                : _unselectedForegroundColor)),
-                    child: tab.icon)
+                    data: IconThemeData(size: 24.0, color: foregroundColor),
+                    child: tab.icon!)
                 : Container(),
             SizedBox(
               width: tab.icon == null || (tab.text == null && tab.child == null)
@@ -352,7 +284,7 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
             ),
             tab.text != null
                 ? Text(
-                    tab.text,
+                    tab.text!,
                     style: textStyle,
                   )
                 : (tab.child ?? Container())
@@ -365,42 +297,44 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
   @override
   Widget build(BuildContext context) {
     assert(() {
-      if (_controller.length != widget.tabs.length) {
+      if (_controller!.length != widget.tabs.length) {
         throw FlutterError(
-            "Controller's length property (${_controller.length}) does not match the "
+            "Controller's length property (${_controller!.length}) does not match the "
             "number of tabs (${widget.tabs.length}) present in TabBar's tabs property.");
       }
       return true;
     }());
-    if (_controller.length == 0) return Container(height: widget.height);
+    if (_controller!.length == 0) return Container(height: widget.height);
 
     _textLTR = Directionality.of(context).index == 1;
     return AnimatedBuilder(
-      animation: _colorTweenBackgroundActivate,
+      animation: _animationController,
       builder: (context, child) => SizedBox(
         key: _tabsContainerKey,
-        height: widget.height,
+        height: widget.preferredSize.height,
+        //height: 58,
         child: ListView.builder(
-            physics: _scrollPhysics,
-            controller: _scrollController,
-            scrollDirection: Axis.horizontal,
-            itemCount: widget.tabs.length,
-            itemBuilder: (BuildContext context, int index) =>
-                _buildButton(index, widget.tabs[index])),
+          physics: widget.physics,
+          controller: _scrollController,
+          scrollDirection: Axis.horizontal,
+          itemCount: widget.tabs.length,
+          itemBuilder: (BuildContext context, int index) =>
+              _buildButton(index, widget.tabs[index] as Tab),
+        ),
       ),
     );
   }
 
   // runs during the switching tabs animation
   _handleTabAnimation() {
-    _aniIndex = ((_controller.animation.value > _prevAniValue)
-            ? _controller.animation.value
+    _aniIndex = ((_controller!.animation!.value > _prevAniValue)
+            ? _controller!.animation!.value
             : _prevAniValue)
         .round();
-    if (!_controller.indexIsChanging && _aniIndex != _currentIndex) {
+    if (!_controller!.indexIsChanging && _aniIndex != _currentIndex) {
       _setCurrentIndex(_aniIndex);
     }
-    _prevAniValue = _controller.animation.value;
+    _prevAniValue = _controller!.animation!.value;
   }
 
   _goToIndex(int index) {
@@ -411,8 +345,8 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
   }
 
   _setCurrentIndex(int index) {
+    // change the index
     setState(() {
-      // change the index
       _prevIndex = _currentIndex;
       _currentIndex = index;
     });
@@ -431,13 +365,14 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
   _scrollTo(int index) {
     // get the screen width. This is used to check if we have an element off screen
     final RenderBox tabsContainer =
-        _tabsContainerKey.currentContext.findRenderObject();
+        _tabsContainerKey.currentContext!.findRenderObject() as RenderBox;
     double screenWidth = tabsContainer.size.width;
 
     // get the button we want to scroll to
-    RenderBox renderBox = _tabKeys[index]?.currentContext?.findRenderObject();
+    RenderBox renderBox =
+        _tabKeys[index].currentContext?.findRenderObject() as RenderBox;
     // get its size
-    double size = renderBox?.size?.width;
+    double size = renderBox.size.width;
     // and position
     double position = renderBox.localToGlobal(Offset.zero).dx;
 
@@ -448,35 +383,31 @@ class _ButtonsTabBarState extends State<ButtonsTabBar>
     if (offset < 0) {
       // get the first button
       renderBox = (_textLTR ? _tabKeys.first : _tabKeys.last)
-          ?.currentContext
-          ?.findRenderObject();
-      if (renderBox != null) {
-        // get the position of the first button of the TabBar
-        position = renderBox.localToGlobal(Offset.zero).dx;
+          .currentContext
+          ?.findRenderObject() as RenderBox;
+      // get the position of the first button of the TabBar
+      position = renderBox.localToGlobal(Offset.zero).dx;
 
-        // if the offset pulls the first button away from the left side, we limit that movement so the first button is stuck to the left side
-        if (position > offset) offset = position;
-      }
+      // if the offset pulls the first button away from the left side, we limit that movement so the first button is stuck to the left side
+      if (position > offset) offset = position;
     } else {
       // if the button is to the right of the middle
 
       // get the last button
       renderBox = (_textLTR ? _tabKeys.last : _tabKeys.first)
-          ?.currentContext
-          ?.findRenderObject();
-      if (renderBox != null) {
-        // get its position
-        position = renderBox.localToGlobal(Offset.zero).dx;
-        // and size
-        size = renderBox.size.width;
+          .currentContext
+          ?.findRenderObject() as RenderBox;
+      // get its position
+      position = renderBox.localToGlobal(Offset.zero).dx;
+      // and size
+      size = renderBox.size.width;
 
-        // if the last button doesn't reach the right side, use it's right side as the limit of the screen for the TabBar
-        if (position + size < screenWidth) screenWidth = position + size;
+      // if the last button doesn't reach the right side, use it's right side as the limit of the screen for the TabBar
+      if (position + size < screenWidth) screenWidth = position + size;
 
-        // if the offset pulls the last button away from the right side limit, we reduce that movement so the last button is stuck to the right side limit
-        if (position + size - offset < screenWidth)
-          offset = position + size - screenWidth;
-      }
+      // if the offset pulls the last button away from the right side limit, we reduce that movement so the last button is stuck to the right side limit
+      if (position + size - offset < screenWidth)
+        offset = position + size - screenWidth;
     }
     offset *= (_textLTR ? 1 : -1);
 
